requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
flask_sqlalchemy
flask_bcrypt
flask_jwt_extended
flask_dotenv
python_dotenv